requirements.txt
autopep8==1.7.0
bcrypt==4.0.1
beanie==1.12.0
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
passlib==1.7.4
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.18.3
